Add get-indexed-image-data
This method on a Png object delivers the image data in
one of two standardized forms: Either a 1-bit bitmap
or an 8 bit indexed image chunk.  This covers indexed
PNGs with 1, 2, 4, or 8 bits per pixel, and also gray
scale PNGs with 8 bits per pixel.  True color and
gray-plus-alpha PNGs are not available through this
API.