Adds PngRandomAccess
This class is intended for use with PNGs produced by
the pngunzip tool in bin/.  Such PNGs may have a compact
representation, like 2 bits per pixel, but they are not
compressed. Therefore we have random access to the pixel
data.  They are designed to be kept in flash and used for
UI elements that are almost always in use.

For UI elements that are large in number, but seldom in use
it may be better to use the PngUncompressed class, which
decompresses to RAM when it is used.